Word insert test correction
Fixes #1
The problem was that the expectation does not expect the characters for
an escaped new line, and the new lines are already there in the test as
represented by a new line.

When we remove the \n characters, then it still fails, because the first
line is not meant to be blank.