Shapes exercise simplified
This modification is very simple.  I have added "puts" in the click
method.  This is appropriate due to the nature of what click does.
The tests break, because what should be tested on a "unit" level is each
thing that click does.  We should trust at this point that click will
perform those actions.

Later, there should be acceptance testing that in fact the click method
does the high order things.

I also removed 'puts' from the running code, as when I click I would see
output on the screen, so I don't need to puts.  In fact, if click does
ever become a graphical thing, rather than textual, I won't need to change
the running code as far as activating a click.